backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.11
django-cors-headers>=4.9.0
djangorestframework
celery>=5.5.3
redis>=7.1.0
python-dotenv>=1.2.1
requests>=2.31.0
openai-whisper
Pillow
google-generativeai
boto3
yt-dlp
opencv-python
stripe